+++ packages/cloud_firestore/cloud_firestore/dartpad/lib/main.dart
@@ -103,7 +103,7 @@ class _FilmListState extends State<FilmList> {
                   style: Theme.of(context).textTheme.bodySmall,
                 );
               },
-            )
+            ),
           ],
         ),
         actions: <Widget>[
@@ -220,7 +220,7 @@ class _MovieItem extends StatelessWidget {
           Likes(
             reference: reference,
             currentLikes: movie.likes,
-          )
+          ),
         ],
       ),
     );
@@ -264,7 +264,7 @@ class _MovieItem extends StatelessWidget {
               style: const TextStyle(color: Colors.white),
             ),
           ),
-        )
+        ),
     ];
   }