+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/lib/src/collection_data.dart
@@ -110,8 +110,9 @@ class CollectionData with Names {
 
     // TODO(validate name)
 
-    final path = annotation.getField('path')!.toStringValue()!;
+    final path = annotation.getField('path')!.toStringValue();
     _assertIsValidCollectionPath(path, annotatedElement);
+    path!;
 
     final type = CollectionData.modelTypeOfAnnotation(annotation);
 
@@ -156,7 +157,7 @@ class CollectionData with Names {
       );
     }
 
-    final annotatedElementSource = annotatedElement.librarySource!;
+    final annotatedElementSource = annotatedElement.librarySource;
 
     // TODO(rrousselGit) handle parts
     // Whether the model class and the reference variable are defined in the same file
@@ -294,8 +295,8 @@ represents the content of the collection must be in the same file.
             ?.contains(classPrefix) ??
         false) {
       throw InvalidGenerationSourceError(
-        'Defined a collection with duplicate class prefix $classPrefix.'
-        ' Either use a different class, or set a unique class prefix.',
+        'Defined a collection with duplicate class prefix $classPrefix. '
+        'Either use a different class, or set a unique class prefix.',
       );
     }
 
@@ -306,12 +307,20 @@ represents the content of the collection must be in the same file.
     return data;
   }
 
-  static void _assertIsValidCollectionPath(String path, Element element) {
+  static void _assertIsValidCollectionPath(String? path, Element element) {
+    if (path == null) {
+      throw InvalidGenerationSourceError(
+        'The annotation @Collection received "$path" as collection path, '
+        'but the path was null, which is not allowed.',
+        element: element,
+      );
+    }
+
     final allowedCharactersRegex = RegExp(r'^[0-9a-zA-Z/*-_]+$');
     if (!allowedCharactersRegex.hasMatch(path)) {
       throw InvalidGenerationSourceError(
         'The annotation @Collection received "$path" as collection path, '
-        ' but the path contains illegal characters.',
+        'but the path contains illegal characters.',
         element: element,
       );
     }
@@ -320,7 +329,7 @@ represents the content of the collection must be in the same file.
     if (pathSplit.length.isEven) {
       throw InvalidGenerationSourceError(
         'The annotation @Collection received "$path" as collection path, '
-        ' but this path points to a document instead of a collection',
+        'but this path points to a document instead of a collection',
         element: element,
       );
     }
@@ -329,7 +338,7 @@ represents the content of the collection must be in the same file.
       if (pathSplit[i] == '*') {
         throw InvalidGenerationSourceError(
           'The annotation @Collection received "$path" as collection path, '
-          ' but ${pathSplit[i]} is not a valid collection name',
+          'but ${pathSplit[i]} is not a valid collection name',
           element: element,
         );
       }
@@ -346,7 +355,7 @@ represents the content of the collection must be in the same file.
         type.isDartCoreInt ||
         type.isDartCoreDouble ||
         type.isDartCoreBool ||
-        type.isPrimitiveList ||
+        type.isSupportedPrimitiveIterable ||
         type.isJsonDocumentReference ||
         dateTimeChecker.isAssignableFromType(type) ||
         timestampChecker.isAssignableFromType(type) ||
@@ -429,7 +438,10 @@ extension on String {
   }
 }
 
-extension on DartType {
+const _coreListChecker = TypeChecker.fromUrl('dart:core#List');
+const _coreSetChecker = TypeChecker.fromUrl('dart:core#Set');
+
+extension DartTypeExtension on DartType {
   bool get isJsonDocumentReference {
     return element?.librarySource?.uri.scheme == 'package' &&
         const {'cloud_firestore'}
@@ -438,8 +450,12 @@ extension on DartType {
         (this as InterfaceType).typeArguments.single.isDartCoreMap;
   }
 
-  bool get isPrimitiveList {
-    if (!isDartCoreList) return false;
+  bool get isSupportedIterable =>
+      _coreListChecker.isExactlyType(this) ||
+      _coreSetChecker.isExactlyType(this);
+
+  bool get isSupportedPrimitiveIterable {
+    if (!isSupportedIterable) return false;
 
     final generic = (this as InterfaceType).typeArguments.single;