+++ packages/firebase_auth/firebase_auth_platform_interface/lib/src/method_channel/method_channel_multi_factor.dart
@@ -40,7 +40,7 @@ class MethodChannelMultiFactor extends MultiFactorPlatform {
     final _assertion = assertion as MultiFactorAssertion;
 
     if (_assertion.credential is PhoneAuthCredential) {
-      final credential = _assertion.credential as PhoneAuthCredential;
+      final credential = _assertion.credential! as PhoneAuthCredential;
       final verificationId = credential.verificationId;
       final verificationCode = credential.smsCode;
 
@@ -58,6 +58,16 @@ class MethodChannelMultiFactor extends MultiFactorPlatform {
             verificationId: verificationId,
             verificationCode: verificationCode,
           ),
+          displayName,
+        );
+      } catch (e, stack) {
+        convertPlatformException(e, stack);
+      }
+    } else if (_assertion is TotpMultiFactorAssertion) {
+      try {
+        await _api.enrollTotp(
+          pigeonDefault,
+          _assertion.assertionId,
           displayName,
         );
       } catch (e, stack) {
@@ -125,7 +135,7 @@ class MethodChannelMultiFactorResolver extends MultiFactorResolverPlatform {
     final _assertion = assertion as MultiFactorAssertion;
 
     if (_assertion.credential is PhoneAuthCredential) {
-      final credential = _assertion.credential as PhoneAuthCredential;
+      final credential = _assertion.credential! as PhoneAuthCredential;
       final verificationId = credential.verificationId;
       final verificationCode = credential.smsCode;
 
@@ -143,6 +153,22 @@ class MethodChannelMultiFactorResolver extends MultiFactorResolverPlatform {
             verificationId: verificationId,
             verificationCode: verificationCode,
           ),
+          null,
+        );
+
+        MethodChannelUserCredential userCredential =
+            MethodChannelUserCredential(_auth, result);
+
+        return userCredential;
+      } catch (e, stack) {
+        convertPlatformException(e, stack);
+      }
+    } else if (_assertion is TotpMultiFactorAssertion) {
+      try {
+        final result = await _api.resolveSignIn(
+          _resolverId,
+          null,
+          _assertion.assertionId,
         );
 
         MethodChannelUserCredential userCredential =
@@ -166,7 +192,13 @@ class MultiFactorAssertion extends MultiFactorAssertionPlatform {
   MultiFactorAssertion(this.credential) : super();
 
   /// Associated credential to the assertion
-  final AuthCredential credential;
+  final AuthCredential? credential;
+}
+
+class PhoneMultiFactorAssertion extends MultiFactorAssertion {
+  PhoneMultiFactorAssertion(
+    PhoneAuthCredential credential,
+  ) : super(credential);
 }
 
 /// Helper class used to generate PhoneMultiFactorAssertions.
@@ -178,6 +210,104 @@ class MethodChannelPhoneMultiFactorGenerator
   MultiFactorAssertionPlatform getAssertion(
     PhoneAuthCredential credential,
   ) {
-    return MultiFactorAssertion(credential);
+    return PhoneMultiFactorAssertion(credential);
+  }
+}
+
+class TotpMultiFactorAssertion extends MultiFactorAssertion {
+  TotpMultiFactorAssertion(
+    this.assertionId,
+  ) : super(null);
+
+  final String assertionId;
+}
+
+/// Helper class used to generate PhoneMultiFactorAssertions.
+class MethodChannelTotpMultiFactorGenerator
+    extends TotpMultiFactorGeneratorPlatform {
+  final _api = MultiFactorTotpHostApi();
+
+  /// Transforms a PhoneAuthCredential into a [MultiFactorAssertion]
+  /// which can be used to confirm ownership of a phone second factor.
+  @override
+  Future<TotpSecretPlatform> generateSecret(
+    MultiFactorSession session,
+  ) async {
+    final pigeonSecret = await _api.generateSecret(session.id);
+    return MethodChannelTotpSecret(
+      pigeonSecret.codeIntervalSeconds,
+      pigeonSecret.codeLength,
+      pigeonSecret.enrollmentCompletionDeadline != null
+          ? DateTime.fromMillisecondsSinceEpoch(
+              pigeonSecret.enrollmentCompletionDeadline!,
+            )
+          : null,
+      pigeonSecret.hashingAlgorithm,
+      pigeonSecret.secretKey,
+    );
+  }
+
+  /// Get a [MultiFactorAssertion]
+  /// which can be used to confirm ownership of a TOTP second factor.
+  @override
+  Future<MultiFactorAssertionPlatform> getAssertionForEnrollment(
+    TotpSecretPlatform secret,
+    String oneTimePassword,
+  ) async {
+    final totpAssertionId =
+        await _api.getAssertionForEnrollment(secret.secretKey, oneTimePassword);
+    return TotpMultiFactorAssertion(totpAssertionId);
+  }
+
+  /// Get a [MultiFactorAssertion]
+  /// which can be used to confirm ownership of a TOTP second factor.
+  @override
+  Future<MultiFactorAssertionPlatform> getAssertionForSignIn(
+    String enrollmentId,
+    String oneTimePassword,
+  ) async {
+    final totpAssertionId =
+        await _api.getAssertionForSignIn(enrollmentId, oneTimePassword);
+    return TotpMultiFactorAssertion(totpAssertionId);
+  }
+}
+
+/// Helper class used to generate PhoneMultiFactorAssertions.
+class MethodChannelTotpSecret extends TotpSecretPlatform {
+  MethodChannelTotpSecret(
+    super.codeIntervalSeconds,
+    super.codeLength,
+    super.enrollmentCompletionDeadline,
+    super.hashingAlgorithm,
+    super.secretKey,
+  );
+
+  final _api = MultiFactorTotpSecretHostApi();
+
+  /// Returns a QR code URL as described in https://github.com/google/google-authenticator/wiki/Key-Uri-Format
+  /// This can be displayed to the user as a QR code to be scanned into a TOTP app like Google Authenticator.
+  /// If the optional parameters are unspecified, an accountName of and issuer of are used.
+  @override
+  Future<String> generateQrCodeUrl({
+    String? accountName,
+    String? issuer,
+  }) async {
+    final pigeonResponse = await _api.generateQrCodeUrl(
+      secretKey,
+      accountName,
+      issuer,
+    );
+    return pigeonResponse;
+  }
+
+  /// Opens the specified QR Code URL in a password manager like iCloud Keychain.
+  @override
+  Future<void> openInOtpApp(
+    String qrCodeUrl,
+  ) async {
+    await _api.openInOtpApp(
+      secretKey,
+      qrCodeUrl,
+    );
   }
 }